backend/requirements.txt
# ============================================================================
# MarketLens Backend -- Pinned Dependencies
# Python 3.11+ required
# Last updated: 2025-01-15
# ============================================================================

# --- Web Framework -----------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- AI / LLM Clients -------------------------------------------------------
anthropic==0.42.0
openai==1.58.1

# --- Data & Analysis ---------------------------------------------------------
pandas==2.3.2
pandas-ta==0.4.71b0
numpy==2.2.6

# --- Charting & Visualization ------------------------------------------------
mplfinance==0.12.10b0
matplotlib==3.9.4

# --- Templating & PDF --------------------------------------------------------
jinja2==3.1.5
fpdf2==2.8.2

# --- Data Validation ---------------------------------------------------------
pydantic==2.10.4
pydantic-settings==2.7.1

# --- Environment & Config ----------------------------------------------------
python-dotenv==1.0.1

# --- Database ----------------------------------------------------------------
aiosqlite==0.20.0

# --- HTTP Client (async) -----------------------------------------------------
httpx==0.28.1

# --- SSE (Server-Sent Events) ------------------------------------------------
sse-starlette==2.2.1

# --- Testing -----------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
